+++ Step_on_step/App/schedule-app/package.json
@@ -22,6 +22,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "backend": "json-server --watch db.json --port 5000",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
