+++ Step_on_step/App/schedule-app/src/components/actions/createContact.js
@@ -1,0 +1,7 @@
+import Contacts from "../../sql/database";
+
+const createContact = (props) => {
+  return Contacts.push(props);
+};
+
+export default createContact;
