+++ Step_on_step/App/schedule-app/src/components/actions/deleteContact.js
@@ -1,0 +1,6 @@
+import Contacts from "../../sql/database";
+
+const deleteContact = (index) => {
+  Contacts.splice(index, 1);
+};
+export default deleteContact;
