--- Step_on_step/App/schedule-app/src/components/createContact.js
@@ -1,7 +1,0 @@
-import contacts from "./sql/database";
-
-const createContact = (newUser) => {
-  contacts.push(newUser);
-};
-
-export default createContact;
