--- Step_on_step/App/schedule-app/src/components/deleteContact.js
@@ -1,6 +1,0 @@
-import contacts from "./sql/database";
-
-const deleteContact = (index) => {
-  contacts.splice(index, 1);
-};
-export default deleteContact;
