--- Step_on_step/App/schedule-app/src/components/sql/database.js
@@ -1,15 +1,0 @@
-const contacts = [
-  {
-    id: "1",
-    name: "Lukita",
-    phone: "21717171717",
-    email: "lukitadagalera@piraca.com",
-  },
-  {
-    id: "2",
-    name: "Sabrina de Mendonça Silva",
-    phone: "21974770964",
-    email: "sms.sabrina1002@gmail.com",
-  },
-];
-export default contacts;
